+++ CHANGELOG.md
@@ -1,4 +1,17 @@
 
+# 2.0.0
+
+Improvements:
+
+  - add support for injecting dot evn files via `__PA_SETTINGS_DOT_ENVS`
+  - automatically load watchfiles if dependency available
+  
+Breaking changes:
+
+  - The HotReloadMixin is deprecated. This fucntionality has been
+    moved into the base class itself, and is auto-loaded if the
+    `watchfiles` module is available
+  
 # 1.2.0
 
 Improvements: